--- a/crates/base/src/macros/test_macros.rs
+++ b/crates/base/src/macros/test_macros.rs
@@ -1,11 +1,11 @@
 #[macro_export]
 macro_rules! integration_test {
     ($main_file:expr, $port:expr, $url:expr, ($($function:tt)+) $(, $termination_token: expr)?) => {
-        let (tx, mut rx) = tokio::sync::mpsc::channel::<base::server::ServerCodes>(1);
+        let (tx, mut rx) = tokio::sync::mpsc::channel::<base::server::ServerHealth>(1);
 
         let signal = tokio::spawn(async move {
-            while let Some(base::server::ServerCodes::Listening) = rx.recv().await {
-                integration_test!(@req $port, $url, ($($function)+));
+            while let Some(base::server::ServerHealth::Listening(event_rx)) = rx.recv().await {
+                integration_test!(@req event_rx, $port, $url, ($($function)+));
             }
             None
         });
@@ -47,11 +47,11 @@ macro_rules! integration_test {
         None
     };
 
-    (@req $port:expr, $url:expr, ($req:expr, $_:expr)) => {
-        return $req($port, $url).await;
+    (@req $event_rx:ident, $port:expr, $url:expr, ($req:expr, $_:expr)) => {
+        return $req($port, $url, $event_rx).await;
     };
 
-    (@req $port:expr, $url:expr, $_:expr) => {
+    (@req $_:ident, $port:expr, $url:expr, $__:expr) => {
         let req = reqwest::get(format!("http://localhost:{}/{}", $port, $url)).await;
         return Some(req);
     };