--- a/crates/base/src/rt_worker/mod.rs
+++ b/crates/base/src/rt_worker/mod.rs
@@ -1,4 +1,5 @@
 pub mod implementation;
+pub mod rt;
 pub mod supervisor;
 pub mod utils;
 pub mod worker;