## What kind of change does this PR introduce?

Refactor

## Description

This is a continuation of #241 .

The PR refactors the part that spawns threads for the isolate into Tokio-based lightweight threads[^1].

In my observation, the PR results in [significant improvements](https://github.com/supabase/edge-runtime/pull/244#issuecomment-1894061954) in terms of memory consumption and request throughput compared to v1.30.0.

However, the memory leak issue will continue to exist as long as #240 is not merged, so expect OOM to be unavoidable over time.

Minor changes
----

* PR improves the stability of `cpu_timer` crate. This resulted in running tests being more stabilized.
* Added tests for various request failure scenarios.
* A request cancellation error by the supervisor can be distinguished clearly as a `WorkerRequestCancelled` error instead of an `InvalidWorkerResponse` error. It is only raised when the script execution reaches the CPU/Wall-clock/Memory limit.[^2]



<img src="https://github.com/supabase/edge-runtime/assets/46702285/5ab2b08b-6bd4-40e5-b15f-05b8664f2c1b" width="32px" />

[^1]: The thread pool size for the isolates can be adjusted through `EDGE_RUNTIME_WORKER_POOL_SIZE` environment variable. The default thread pool size follows the return value of `std::thread::available_parallelism`.

[^2]: If it is too small to maintain the isolation in stable, It can raise various errors including the `WorkerRequestCancelled`.