@nyannyacha 2024-01-15
As I said in the conversation of the previous PR, I've modified integration tests to be compatible with the new policies.

Especially, I've made changes for the tests to be run on the `oneshot` policy that's the same as the `per_request` policy but guarantees the exit of the isolate after the request.