@nyannyacha 2024-01-16
Characteristics `per_worker` (v1.30.0 vs PR-244) / 3 min[^1]
----
Running environment
```
Hardware: Macbook Air M1 2020, 16GB
Container: Docker/Colima(0.6.7, vz, no rosetta 2)
Guest OS: Debian/Bookworm
ARCH: aarch64/arm64
VCPU: 8
RAM: 8GiB (16GiB Swap)
```

Limits
```
const workerTimeoutMs = 20 * 1000;
const cpuTimeSoftLimitMs = 500;
const cpuTimeHardLimitMs = 600;
```

Command (v1.30.0)
```
start --main-service /home/deno/functions/main -p 8888
```

Command (PR-244)
```
start --main-service /home/deno/functions/main --max-parallelism=20 --request-wait-timeout=1000000 -p 8888
```

Command
```
vegeta attack -rate=0 -duration=3m -max-workers=12
```

Max-RSS
----

### v1.30.0

Before
<img width="1439" alt="before v130" src="https://github.com/supabase/edge-runtime/assets/46702285/70546d05-fb16-4681-ab53-cd657d0c2da0">

Middle (Max)
<img width="1439" alt="middle v130" src="https://github.com/supabase/edge-runtime/assets/46702285/8d8854bc-e474-4149-8fb2-eb1c5c30cfbb">

After
<img width="1439" alt="after v130" src="https://github.com/supabase/edge-runtime/assets/46702285/db110ad3-ea90-46ed-a665-04ae01a82725">

Report
```
Requests      [total, rate, throughput]         335246, 1862.32, 1857.90
Duration      [total, attack, wait]             3m0s, 3m0s, 178.789ms
Latencies     [min, mean, 50, 90, 95, 99, max]  415.111µs, 5.345ms, 1.484ms, 4.533ms, 5.965ms, 134.48ms, 1.734s
Bytes In      [total, mean]                     16435999, 49.03
Bytes Out     [total, mean]                     7375412, 22.00
Success       [ratio]                           99.86%
Status Codes  [code:count]                      200:334783  500:463
Error Set:
500 Internal Server Error
```

### PR-244

Before
<img width="1439" alt="before meow" src="https://github.com/supabase/edge-runtime/assets/46702285/5b1a24b3-76f6-4ede-9f09-1cb50cfab1c9">

Middle (Max)
<img width="1439" alt="middle meow" src="https://github.com/supabase/edge-runtime/assets/46702285/2f9e6290-9e14-43e8-a94b-7232566202dd">

After
<img width="1439" alt="after meow" src="https://github.com/supabase/edge-runtime/assets/46702285/a94c4b9a-9900-4a8e-b3cf-e18bcc913e27">

Report
```
Requests      [total, rate, throughput]         638162, 3545.35, 3545.33
Duration      [total, attack, wait]             3m0s, 3m0s, 978.813µs
Latencies     [min, mean, 50, 90, 95, 99, max]  500.949µs, 2.143ms, 1.627ms, 4.188ms, 5.046ms, 7.096ms, 125.404ms
Bytes In      [total, mean]                     31269938, 49.00
Bytes Out     [total, mean]                     14039564, 22.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:638162
Error Set:
```


<img src="https://github.com/supabase/edge-runtime/assets/46702285/5ab2b08b-6bd4-40e5-b15f-05b8664f2c1b" width="32px" />


[^1]: benchmark was run in the same host, so the request per sec may not be accurate